ssv-ibft-test- renaming folder to spec_testing
