+++ ibft/change_round_test.go
@@ -10,7 +10,7 @@ import (
 	msgcontinmem "github.com/bloxapp/ssv/ibft/msgcont/inmem"
 	"github.com/bloxapp/ssv/ibft/pipeline/changeround"
 	"github.com/bloxapp/ssv/ibft/proto"
-	ibfttesting "github.com/bloxapp/ssv/ibft/testing"
+	ibfttesting "github.com/bloxapp/ssv/ibft/spec_testing"
 )
 
 func changeRoundDataToBytes(input *proto.ChangeRoundData) []byte {
