+++ ibft/instance_test.go
@@ -7,7 +7,7 @@ import (
 	"go.uber.org/zap/zaptest"
 
 	"github.com/bloxapp/ssv/ibft/proto"
-	ibfttesting "github.com/bloxapp/ssv/ibft/testing"
+	ibfttesting "github.com/bloxapp/ssv/ibft/spec_testing"
 	"github.com/bloxapp/ssv/network/local"
 	"github.com/bloxapp/ssv/utils/dataval/bytesval"
 )
