+++ ibft/prepare_test.go
@@ -9,7 +9,7 @@ import (
 
 	msgcontinmem "github.com/bloxapp/ssv/ibft/msgcont/inmem"
 	"github.com/bloxapp/ssv/ibft/proto"
-	ibfttesting "github.com/bloxapp/ssv/ibft/testing"
+	ibfttesting "github.com/bloxapp/ssv/ibft/spec_testing"
 	"github.com/bloxapp/ssv/utils/dataval/bytesval"
 )
